Setup the virtualenv
As Henrique mentioned, we will most likely need a virtual env for this to run easily the more we add dependencies and what not. So I created just a basic environment with:

Python 3.9
Django

postgresql does not need to be in this environment as it is not a python module.